Request various details as attributes on Confirm.
Confirm requires the SiteCode to be set when making NewEnquiries, so
we publish this as a metadata attribute on all services. Confirm also
has individual fields for the enquiry location/title & description and
asset ID, so these are included also.

This commit makes it simple for integrations to use their own Service
classes, if they have specific metadata requirements.