[Confirm] Use LoggedTime instead of LogEffectiveTime for StatusLogNotes
LoggedTime is the timestamp that the StatusLog was added to the Enquiry;
LogEffectiveTime is the timestamp when that StatusLog was created.

There is a subtle difference here, which was causing problems when
Enquiries were linked together within Confirm and StatusLogNotes
transferred from one Enquiry to another (e.g. when multiple Enquiries
were assigned to the same job). In this situation the LogEffectiveTime
of the most recently-added StatusLog may have an earlier
LogEffectiveTime than other StatusLogs on the Enquiry, or even the
Enquiry itself.

Although open311-adapter was doing the right thing and presenting all
StatusLog entries in the servicerequestupdates.xml output, those with
an earlier LogEffectiveTime were being ignored by FixMyStreet meaning
those reports weren't being updated appropriately.

Fixes mysociety/fixmystreet-commercial#1102.