[Alloy] handle true/false contact defaults
For true/false values we have to use the JSON()->true/false otherwise
when we convert to JSON later we get 1/0 which then fails the validation
at the Alloy end
NB: have to use 'true'/'false' strings in the YAML for this to work. If we
use true/false then it gets passed in as something that gets converted to
1/0

We could possibly use local $YAML::XS::Boolean = "JSON::PP" in the Config
module to get round all this but not sure if that would break something
else.