[Alloy] update fetching
In Alloy once the initial inspection has happened a defect is created
from the inspection and the inspection closed. Because of this we need
to get updated inspections and updated defects in order to follow
through the entire cycle of a report.

Linking defects and inspection is usually done by looking for an
inspection linked to the defect, however in certain cases it's not
possibly to create this link so the defect will have a fixmystreet id
attribute which we pass to FMS as fixmystreet_id so it can use this in
preference to the external_id when matching.