[Alloy] report fetching
Fetch updated defects and if they aren't linked to an inspection, or
have a fixmystreet id then import them. There's a fairly complicated
defect type to category mapping scheme because defects have types, which
are relevant to the category we map them to.

Also worth noting that sometimes defects have a lineString geometry and
at the moment we are picking the co-ordinates half way through the array
and using those as the location. This is not ideal but mostly defects
should have a point geometry.