[Northamptonshire] ignore linked defects until scheduled
The flow from inspection to defect goes:
1: inspection created
2: inspection assigned -> FMS to investigating
3: inspection closed, set to works instructed -> FMS action scheduled
4: defect created
5: defect pulled in to Keir, defect set to order raised
6: Fixed, defect to remedied -> FMS fixed

At step 4 the defect is confirmed which maps to investigating. We want
to ignore this step for defects linked to inspections as othwise the FMS
status goes

investigating
action scheduled
investigating
action scheduled

which is confusing.

However, we want non linked defects to follow a normal course.