[Confirm] Photo upload.
The process of creating a new enquiry and then uploading its photos
requires multiple API calls so is not an atomic operation.
Confirm can be slow to upload photos after we've created an enquiry.
If it's *really* slow the FMS POST Service Request will timeout and FMS
will make another attempt at sending the report later on, which is a
problem because the Enquiry has already been raised so we end up with
duplicate Enquiries in Confirm.

This change uses a much more aggressive timeout for photo uploads, and
ensures that timeouts/other errors that may occur don't stop FMS from
considering the report sent succcessfully.