[Alloy] use current time for comment time
There is a race condition on the initial update to Alloy problems where
the response back to FMS from creating the report happens after Alloy
has updated the state of the report in Alloy.

This means that the time of the comment is before the last update time
of the report in FMS which means we don't update the state of the report
and you can end up with two comments setting the report to
Investigating.

By setting the time of the comment to now -20 seconds we get round this.