[Northamptonshire] always set external_status_code on defect updates
We need to set this to stop phantom updates being produced. This
happens because when an inspection is closed it always sets an
external_status_code which we never unset. Then when updates arrive
from defects with no external_status_code the template fetching code at
FixMyStreet sees that the external_status_code has changed and fetches
the template. This means we always get an update even if nothing has
changed. So, set the external_status_code on defects associated with an
inspection to the external_status_code used when an inspection is marked
for raising as a defect.