Fix issue where configuration lost calling proxy.
I haven't worked out the details as to how this is happening, but a call
to $self->proxy always calls new() and creates a new object, which isn't
returned, but somehow ends up blanking out the config_filename then used
to lookup the server config.

If the config was looked up at all before calling proxy, it was then not
affected, so this turned out to only affect Symology as the Confirm code
always looks up stuff first.

I have 'fixed' this by removing the check for proxy being set, as we can
always set it to the configured variable.