[Ezytreev] Use new StatusHistory field for full updates
Previously we only got the latest update for an enquiry, which meant if
someone had changed the status multiple times before closing the enquiry
modal in ezytreev we'd only get the latest status.

Ezytreev have now added a "StatusHistory" field to the API which
contains a full list of updates, along with a unique ID for each one, so
we can use that to ensure we don't miss any updates.