[Ezytreev] Leave description blank by default
In FMS automated response templates are only used if the description
from open311 is blank or the response template contains the text
`{{description}}`, which is then replaced with the actual description
from open311. In this case the ezytreev description is just the ezytreev
internal status description, which isn't relevant to the end user, and
could cause confusion since the wording is different from FMS statuses,
so leave it blank and let the automated response template take priority.