[Alloy] fetching of completion photos
This will fetch any photos that were added as part of an update with a
status set in the completion_photo_states config option. If this config
option is missing then it skips fetching photos.

This also adds the /photos/completion endpoint to the Alloy integration
for fetching photos from Alloy.