[Northamptonshire] Use narrower search criteria when looking up contact
The GlobalAttributeSearch was returning results for contacts whose email addresses
only had a partial match to the search term. Because we were just taking the first
result, this sometimes resulted in the wrong contact being associated with the
report in Alloy.

This commit uses a stricter comparison operator to search for contacts whose email
or phone exactly match the search term. It also introduces our own equivalence check
to make sure we’re definitely looking at the right contact.