[Confirm] Remove file locking.
This data is currently held over NFS and we were seeing stuck flock calls.
These files are not written to, only read or removed, and the script only
runs on one server anyway, it would still have a race condition if run on
more than one even with shared locking.