Allow any service attribute through by default
Previously this was only for the Echo integration, but this is
probably a sensible default generally, in the spirit of being liberal
with what you receive.

I'm printing out a warning to the logfile if an attribute is
unrecognised, but these are suppressed if the service has the
allow_any_attributes flag set, which is just Echo for now.