[Bartec] Handle extended data values without ‘ - ’ correctly
Not all Bartec values follow the pattern the regex was looking for. As a result
the $1 might have been empty causing a schema validation error when trying to
fetch the service metadata.

Additionally, on macOS the $1 value was sometimes ‘darwin’, oddly, which meant
that all values had that as their label if the regex didn’t match.