[Oxfordshire][Open311] Finer control for jobs with parent defects updating FMS
Issue with Oxfordshire having jobs within Alloy updating FMS statuses
when the job has a parent defect.

They only want the status on FMS to be updated when the parent status is updated,
which is what they want to be reflected to the citizen.

This adds finer control over which type of jobs with parent defects should be skipped over
by respective councils.

Maintains the status quo for all except Oxford.

https://mysocietysupport.freshdesk.com/a/tickets/2668