[Abavus] Add adding of reference and systemCode
When creating a report it needs to have a systemCode and
a reference added - systemCode is hardcoded to 'FMS'
to match the Bucks profile.

reference is the FMS id as the external id.

This is for the forthcoming 'event' call that will allow
fetching updated reports and will need these present
to work at the Abavus end.

https://3.basecamp.com/4020879/buckets/32325686/todos/6191167800#__recording_6245727608