[Alloy] Improve service code usage.
Assign unique codes to categories with an alias, and spot categories
with the same name in multiple groups; use the group passed in directly
where necessary (Northumberland). This should allow the use of
categories with the same name in multiple groups without having to have
different names (meaning most aliases can be removed from configuration)
and allow two different categories to map to the same Alloy category.