[Confirm] Fetch defects
Defects differ from jobs in that they don't receive
updates directly on them in Confirm. Instead, an
attached job receives an update, which we use to
determine the Open311 status of the service request.

Service requests that originate as defects can be
published under an existing Open311 service, or
under a defect-specific service, according to the
configuration of defect_service_mapping.

For https://github.com/mysociety/societyworks/issues/4899