[Hampshire][Confirm] Adds an allowed_attributes attribute
When this exists in config, it works to only allow attributes
included in the list. It is used in place of ignore_attributes
which will be discounted if it also exists.

If there is an empty list in allow_attributes, it will mean
that no attributes will be passed through.

https://github.com/mysociety/societyworks/issues/5102